refactor(server): extract reconcile business logic to internal/reconcile/
Split reconcile.go (~1318 lines) into:
- internal/reconcile/reconcile.go — all types + pure-logic functions
  (RunReconcileScan, BuildReconcilePreview, ExecuteReconcile, etc.)
- internal/server/reconcile.go — HTTP handlers only (~203 lines)

reconcile.Store exported as package interface. Pure-logic functions
converted from *Server receivers to free functions with explicit store param.
Updated scheduler.go and server_lifecycle.go to call reconcile.* functions.
Updated reconcile_test.go to use exported reconcile package functions.

Refs: PKG-3
