+++ internal/server/reconcile_test.go
@@ -1,11 +1,13 @@
 // file: internal/server/reconcile_test.go
-// version: 1.0.0
+// version: 2.0.0
 // guid: f8a9b0c1-d2e3-4f5a-6b7c-8d9e0f1a2b3c
 
 package server
 
 import (
 	"testing"
+
+	"github.com/jdfalk/audiobook-organizer/internal/reconcile"
 )
 
 func TestNormalizeFilename(t *testing.T) {
@@ -23,31 +25,31 @@ func TestNormalizeFilename(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		result := normalizeFilename(tt.input)
+		result := reconcile.NormalizeFilename(tt.input)
 		if result != tt.expected {
-			t.Errorf("normalizeFilename(%q) = %q, want %q", tt.input, result, tt.expected)
+			t.Errorf("NormalizeFilename(%q) = %q, want %q", tt.input, result, tt.expected)
 		}
 	}
 }
 
 func TestCountMatchType(t *testing.T) {
-	matches := []ReconcileMatch{
+	matches := []reconcile.ReconcileMatch{
 		{MatchType: "hash"},
 		{MatchType: "hash"},
 		{MatchType: "filename"},
 		{MatchType: "original_hash"},
 	}
 
-	if got := countMatchType(matches, "hash"); got != 2 {
-		t.Errorf("countMatchType(hash) = %d, want 2", got)
+	if got := reconcile.CountMatchType(matches, "hash"); got != 2 {
+		t.Errorf("CountMatchType(hash) = %d, want 2", got)
 	}
-	if got := countMatchType(matches, "filename"); got != 1 {
-		t.Errorf("countMatchType(filename) = %d, want 1", got)
+	if got := reconcile.CountMatchType(matches, "filename"); got != 1 {
+		t.Errorf("CountMatchType(filename) = %d, want 1", got)
 	}
-	if got := countMatchType(matches, "original_hash"); got != 1 {
-		t.Errorf("countMatchType(original_hash) = %d, want 1", got)
+	if got := reconcile.CountMatchType(matches, "original_hash"); got != 1 {
+		t.Errorf("CountMatchType(original_hash) = %d, want 1", got)
 	}
-	if got := countMatchType(matches, "nonexistent"); got != 0 {
-		t.Errorf("countMatchType(nonexistent) = %d, want 0", got)
+	if got := reconcile.CountMatchType(matches, "nonexistent"); got != 0 {
+		t.Errorf("CountMatchType(nonexistent) = %d, want 0", got)
 	}
 }
